Update hero
The hero in the original design left something to be desired. I created a new one that should give a much more professional and modern touch to TEAFC

Update Hero

The hero in the original design left something to be desired. I created a new one that should give a much more professional and modern touch to TEAFC